Implementando jsp parcialmente funcional
segunda parte do jsp será commitada em duas 3 horas

+++ src/java/br/pd/inf/sistemahospital/service/Implementacao/KeepCadastroImplementacao.java
@@ -61,7 +61,7 @@ public class KeepCadastroImplementacao implements KeepCadastro{
                 this.ficha.edit(ficha);
             }
                 @Override
-                public FichaPaciente pesquisarFichaByID(long Id){
+                public FichaPaciente pesquisarFichaByID(int Id){
                     
                     FichaPaciente resultado = (FichaPaciente) ficha.find(Id);
                     return resultado;
